client/requirements.txt
streamlit==1.32.0
mysql-connector-python==8.3.0